Add support for deploying Monasca Thresh
Monasca Thresh is a Java topology for Storm which generates
alerts from metrics. It depends on Java code in the
monasca-common library. In this change we support building
the Monasca Thresh Storm topology from source. To enable us
to submit the topology to Storm we inherit from the Storm
container, solely for the Storm client.

Change-Id: I389af89086cdab6d243c4a4fa65fd157535a5045
Partially-Implements: blueprint monasca-containers

Conflicts:
	kolla/common/config.py