Add Monasca Agent container
The Monasca Agent consists of 3 processes. The collector, which
collects system level metrics from a host, a statsd server, which is used
to recieve application specific statsd metrics and a forwarder service
which batches metrics and sends them to the Monasca API at pre-defined
time intervals.

Partially-Implements: blueprint monasca-containers
Change-Id: Ia446ca0a6b932c2429d727b63e363b4e27307304

Conflicts:
	kolla/common/config.py