Build RabbitMQ 4.1 in stable releases
RabbitMQ 4.1 is released, meaning 4.0 and prior are out of community
support. This patch allow for building 4.1 in the `rabbitmq-4-1`
container, while keeping the existing version in the standard `rabbitmq`
container.

We do not need to build 4.0 in past releases, as 4.1 supports upgrades
from 4.0.x and 3.13.x. [1].

Same method as in: https://review.opendev.org/c/openstack/kolla/+/918974

1. https://github.com/rabbitmq/rabbitmq-server/releases/tag/v4.1.0

Change-Id: I4a4fe7244ea0b92cf12b96f4dee8a7abd34028d9